Fix bug that compiled using wrong params.h
The params.h file was not actually adjusted to point to the right
parameter header for a specific parameter set. This is somewhat
non-trivial because FFI objects are collected much earlier than
the actual compilation. We resolve this by creating directories
for each of the different instances with complete copies of the
reference source code.